+++ src/main/java/com/avrix/ui/Widget.java
@@ -12,29 +12,12 @@ import static org.lwjgl.nanovg.NanoVG.*;
  * for UI elements that can be rendered and interacted with.
  */
 public abstract class Widget {
-    /**
-     * Indicates whether this {@link Widget} is visible or not.
-     */
     private boolean visible = true;
+    private boolean alwaysOnTop = false;
 
-    /**
-     * The x-coordinate of the top-left corner of this {@link Widget}.
-     */
     private int x;
-
-    /**
-     * The y-coordinate of the top-left corner of this {@link Widget}.
-     */
     private int y;
-
-    /**
-     * The width of this {@link Widget}.
-     */
     private int width;
-
-    /**
-     * The height of this {@link Widget}.
-     */
     private int height;
 
     private UIContext context;
@@ -57,89 +40,162 @@ public abstract class Widget {
     /**
      * Called when the mouse is moved over the {@link Widget}.
      *
-     * @param context the UI context in which the event occurred
-     * @param x       the x-coordinate of the mouse position
-     * @param y       the y-coordinate of the mouse position
+     * @param x relative x-coordinate of mouse position
+     * @param y relative y-coordinate of the mouse position
      */
-    public void onMouseMove(UIContext context, float x, float y) {
+    public void onMouseMove(int x, int y) {
     }
 
     /**
      * Called when the left mouse button is pressed down over the {@link Widget}.
      *
-     * @param context the UI context in which the event occurred
-     * @param x       the x-coordinate of the mouse position
-     * @param y       the y-coordinate of the mouse position
+     * @param x relative x-coordinate of the mouse position
+     * @param y relative y-coordinate of the mouse position
      */
-    public void onLeftMouseDown(UIContext context, float x, float y) {
+    public void onLeftMouseDown(int x, int y) {
+        bringToTop();
     }
 
     /**
      * Called when the left mouse button is released over the {@link Widget}.
      *
-     * @param context the UI context in which the event occurred
-     * @param x       the x-coordinate of the mouse position
-     * @param y       the y-coordinate of the mouse position
+     * @param x relative x-coordinate of the mouse position
+     * @param y relative y-coordinate of the mouse position
      */
-    public void onLeftMouseUp(UIContext context, float x, float y) {
+    public void onLeftMouseUp(int x, int y) {
     }
 
     /**
      * Called when the right mouse button is pressed down over the {@link Widget}.
      *
-     * @param context the UI context in which the event occurred
-     * @param x       the x-coordinate of the mouse position
-     * @param y       the y-coordinate of the mouse position
+     * @param x relative x-coordinate of the mouse position
+     * @param y relative y-coordinate of the mouse position
      */
-    public void onRightMouseDown(UIContext context, float x, float y) {
+    public void onRightMouseDown(int x, int y) {
+        bringToTop();
     }
 
     /**
      * Called when the right mouse button is released over the {@link Widget}.
      *
-     * @param context the UI context in which the event occurred
-     * @param x       the x-coordinate of the mouse position
-     * @param y       the y-coordinate of the mouse position
+     * @param x relative x-coordinate of the mouse position
+     * @param y relative y-coordinate of the mouse position
      */
-    public void onRightMouseUp(UIContext context, float x, float y) {
+    public void onRightMouseUp(int x, int y) {
     }
 
     /**
      * Called when the mouse wheel is scrolled over the {@link Widget}.
      *
-     * @param context the UI context in which the event occurred
-     * @param x       the x-coordinate of the mouse position
-     * @param y       the y-coordinate of the mouse position
-     * @param delta   the amount of scrolling; positive values for scrolling up, negative for scrolling down
+     * @param x     relative x-coordinate of the mouse position
+     * @param y     relative y-coordinate of the mouse position
+     * @param delta direction of mouse wheel movement - (1 - up, -1 - down)
      */
-    public void onMouseWheel(UIContext context, float x, float y, float delta) {
+    public void onMouseWheel(int x, int y, int delta) {
+    }
+
+    /**
+     * Handles the mouse move event outside any visible widget
+     *
+     * @param x absolute x-coordinate of the mouse position
+     * @param y absolute y-coordinate of the mouse position
+     */
+    public void onMouseMoveOutside(int x, int y) {
+    }
+
+    /**
+     * Handles the left mouse button down event outside any visible widget
+     *
+     * @param x absolute x-coordinate of the mouse position
+     * @param y absolute y-coordinate of the mouse position
+     */
+    public void onLeftMouseDownOutside(int x, int y) {
+    }
+
+    /**
+     * Handles the left mouse button up event outside any visible widget
+     *
+     * @param x absolute x-coordinate of the mouse position
+     * @param y absolute y-coordinate of the mouse position
+     */
+    public void onLeftMouseUpOutside(int x, int y) {
+    }
+
+    /**
+     * Handles the right mouse button down event outside any visible widget
+     *
+     * @param x absolute x-coordinate of the mouse position
+     * @param y absolute y-coordinate of the mouse position
+     */
+    public void onRightMouseDownOutside(int x, int y) {
+    }
+
+    /**
+     * Handles the right mouse button up event outside any visible widget
+     *
+     * @param x absolute x-coordinate of the mouse position
+     * @param y absolute y-coordinate of the mouse position
+     */
+    public void onRightMouseUpOutside(int x, int y) {
+    }
+
+    /**
+     * Handles the mouse wheel event outside any visible widget
+     *
+     * @param x     absolute x-coordinate of the mouse position
+     * @param y     absolute y-coordinate of the mouse position
+     * @param delta the amount of wheel movement
+     */
+    public void onMouseWheelOutside(int x, int y, int delta) {
     }
 
     /**
      * Called when a key is pressed while the {@link Widget} has focus.
      *
-     * @param context the UI context in which the event occurred
-     * @param key     the code of the key that was pressed
+     * @param key the code of the key that was pressed
      */
-    public void onKeyPress(UIContext context, int key) {
+    public void onKeyPress(int key) {
     }
 
     /**
      * Called when a key is released while the {@link Widget} has focus.
      *
-     * @param context the UI context in which the event occurred
-     * @param key     the code of the key that was released
+     * @param key the code of the key that was released
      */
-    public void onKeyRelease(UIContext context, int key) {
+    public void onKeyRelease(int key) {
     }
 
     /**
      * Called when a key is held down and repeatedly pressed while the {@link Widget} has focus.
      *
-     * @param context the UI context in which the event occurred
-     * @param key     the code of the key that is being repeatedly pressed
+     * @param key the code of the key that is being repeatedly pressed
      */
-    public void onKeyRepeat(UIContext context, int key) {
+    public void onKeyRepeat(int key) {
+    }
+
+    /**
+     * Sets whether the window should always be on top of other windows.
+     *
+     * @param alwaysOnTop {@code true} if the window should always be on top; otherwise {@code false}
+     */
+    public void setAlwaysOnTop(boolean alwaysOnTop) {
+        this.alwaysOnTop = alwaysOnTop;
+    }
+
+    /**
+     * Checks if the window is set to always be on top of other windows.
+     *
+     * @return {@code true} if the window is always on top; otherwise {@code false}
+     */
+    public boolean isAlwaysOnTop() {
+        return this.alwaysOnTop;
+    }
+
+    /**
+     * Moves this {@link Widget} to the front of the rendering order, ensuring it is drawn above other widgets.
+     */
+    public void bringToTop() {
+        WidgetManager.bringWidgetToTop(this);
     }
 
     /**
@@ -259,7 +315,7 @@ public abstract class Widget {
      * Adds this {@link Widget} to the screen by registering it with the {@link WidgetManager}.
      * The {@link Widget} will be managed and rendered as part of the UI.
      */
-    public void addToScreen() {
+    public final void addToScreen() {
         WidgetManager.addWidget(this);
     }
 
@@ -267,7 +323,7 @@ public abstract class Widget {
      * Removes this {@link Widget} from the screen by deregistering it from the {@link WidgetManager}.
      * The {@link Widget} will no longer be managed or rendered as part of the UI.
      */
-    public void removeFromScreen() {
+    public final void removeFromScreen() {
         WidgetManager.removeWidget(this);
     }
 
@@ -276,7 +332,7 @@ public abstract class Widget {
      *
      * @param context the {@link UIContext} to be set
      */
-    public void setContext(UIContext context) {
+    public final void setContext(UIContext context) {
         this.context = context;
     }
 
@@ -285,8 +341,86 @@ public abstract class Widget {
      *
      * @return the {@link UIContext} currently associated with this {@link Widget}
      */
-    public UIContext getContext() {
+    public final UIContext getContext() {
         return this.context;
+    }
+
+    /**
+     * Returns the current {@link UIContext} ID for this {@link Widget}.
+     *
+     * @return the {@link UIContext} ID currently associated with this {@link Widget}
+     */
+    public final long getContextID() {
+        return this.context.get();
+    }
+
+    /**
+     * Sets the scissor region for NanoVG. Only content inside this region will be rendered.
+     *
+     * @param x      the x-coordinate of the scissor region.
+     * @param y      the y-coordinate of the scissor region.
+     * @param width  the width of the scissor region.
+     * @param height the height of the scissor region.
+     */
+    public void startScissor(int x, int y, int width, int height) {
+        nvgScissor(getContextID(), x, y, width, height);
+    }
+
+    /**
+     * Sets the intersecting scissor region for NanoVG. Only content inside this intersection of
+     * this region and the current scissor region will be rendered.
+     *
+     * @param x      the x-coordinate of the intersecting scissor region.
+     * @param y      the y-coordinate of the intersecting scissor region.
+     * @param width  the width of the intersecting scissor region.
+     * @param height the height of the intersecting scissor region.
+     */
+    public void startIntersectScissor(int x, int y, int width, int height) {
+        nvgIntersectScissor(getContextID(), x, y, width, height);
+    }
+
+    /**
+     * Resets the scissor region, allowing rendering to occur over the entire canvas.
+     */
+    public void endScissor() {
+        nvgResetScissor(getContextID());
+    }
+
+    /**
+     * Draws a rectangle with a stroke and no fill
+     *
+     * @param x         the X coordinate of the top left corner of the rectangle
+     * @param y         the Y coordinate of the top left corner of the rectangle
+     * @param width     the width of the rectangle
+     * @param height    height of the rectangle
+     * @param lineWidth the width of the stroke line in pixels
+     * @param color     the color of the outline
+     */
+    public void drawRectOutline(int x, int y, int width, int height, float lineWidth, UIColor color) {
+        nvgBeginPath(getContextID());
+        nvgRect(getContextID(), x, y, width, height);
+        color.tallocNVG(nvgColor -> nvgStrokeColor(getContextID(), nvgColor));
+        nvgStrokeWidth(getContextID(), lineWidth);
+        nvgStroke(getContextID());
+    }
+
+    /**
+     * Draws a rectangle with rounded corners and a stroke without fill.
+     *
+     * @param x         the X coordinate of the top left corner of the rectangle
+     * @param y         the Y coordinate of the top left corner of the rectangle
+     * @param width     the width of the rectangle
+     * @param height    height of the rectangle
+     * @param radius    corner radius
+     * @param lineWidth the width of the stroke line in pixels
+     * @param color     the color of the outline
+     */
+    public void drawRoundedRectOutline(int x, int y, int width, int height, int radius, float lineWidth, UIColor color) {
+        nvgBeginPath(getContextID());
+        nvgRoundedRect(getContextID(), x, y, width, height, radius);
+        color.tallocNVG(nvgColor -> nvgStrokeColor(getContextID(), nvgColor));
+        nvgStrokeWidth(getContextID(), lineWidth);
+        nvgStroke(getContextID());
     }
 
     /**
@@ -299,12 +433,10 @@ public abstract class Widget {
      * @param color  the color to fill the rectangle with
      */
     public void drawRect(int x, int y, int width, int height, UIColor color) {
-        nvgBeginPath(context.get());
-        nvgRect(context.get(), x, y, width, height);
-        color.tallocNVG(nvgColor -> {
-            nvgFillColor(context.get(), nvgColor);
-        });
-        nvgFill(context.get());
+        nvgBeginPath(getContextID());
+        nvgRect(getContextID(), x, y, width, height);
+        color.tallocNVG(nvgColor -> nvgFillColor(getContextID(), nvgColor));
+        nvgFill(getContextID());
     }
 
     /**
@@ -318,13 +450,13 @@ public abstract class Widget {
      * @param color    the color of the text
      */
     public void drawText(String text, String fontName, int x, int y, int fontSize, UIColor color) {
-        nvgFontFace(context.get(), fontName);
-        nvgFontSize(context.get(), fontSize);
-        nvgBeginPath(context.get());
-        color.tallocNVG(nvgColor -> nvgFillColor(context.get(), nvgColor));
-        nvgText(context.get(), x, y, text);
-        nvgFill(context.get());
-        nvgClosePath(context.get());
+        nvgFontFace(getContextID(), fontName);
+        nvgFontSize(getContextID(), fontSize);
+        nvgBeginPath(getContextID());
+        color.tallocNVG(nvgColor -> nvgFillColor(getContextID(), nvgColor));
+        nvgText(getContextID(), x, y, text);
+        nvgFill(getContextID());
+        nvgClosePath(getContextID());
     }
 
     /**
@@ -338,12 +470,10 @@ public abstract class Widget {
      * @param color  the color to fill the rectangle with
      */
     public void drawRoundedRect(int x, int y, int width, int height, float radius, UIColor color) {
-        nvgBeginPath(context.get());
-        nvgRoundedRect(context.get(), x, y, width, height, radius);
-        color.tallocNVG(nvgColor -> {
-            nvgFillColor(context.get(), nvgColor);
-        });
-        nvgFill(context.get());
+        nvgBeginPath(getContextID());
+        nvgRoundedRect(getContextID(), x, y, width, height, radius);
+        color.tallocNVG(nvgColor -> nvgFillColor(getContextID(), nvgColor));
+        nvgFill(getContextID());
     }
 
     /**
@@ -356,12 +486,10 @@ public abstract class Widget {
      * @param color  the color of the ellipse
      */
     public void drawEllipse(int x, int y, int width, int height, UIColor color) {
-        nvgBeginPath(context.get());
-        nvgEllipse(context.get(), x, y, (float) width / 2, (float) height / 2);
-        color.tallocNVG(nvgColor -> {
-            nvgFillColor(context.get(), nvgColor);
-        });
-        nvgFill(context.get());
+        nvgBeginPath(getContextID());
+        nvgEllipse(getContextID(), x, y, (float) width / 2, (float) height / 2);
+        color.tallocNVG(nvgColor -> nvgFillColor(getContextID(), nvgColor));
+        nvgFill(getContextID());
     }
 
     /**
@@ -375,14 +503,12 @@ public abstract class Widget {
      * @param color the color of the line
      */
     public void drawLine(int x1, int y1, int x2, int y2, float width, UIColor color) {
-        nvgBeginPath(context.get());
-        nvgMoveTo(context.get(), x1, y1);
-        nvgLineTo(context.get(), x2, y2);
-        nvgStrokeWidth(context.get(), width);
-        color.tallocNVG(nvgColor -> {
-            nvgStrokeColor(context.get(), nvgColor);
-        });
-        nvgStroke(context.get());
+        nvgBeginPath(getContextID());
+        nvgMoveTo(getContextID(), x1, y1);
+        nvgLineTo(getContextID(), x2, y2);
+        nvgStrokeWidth(getContextID(), width);
+        color.tallocNVG(nvgColor -> nvgStrokeColor(getContextID(), nvgColor));
+        nvgStroke(getContextID());
     }
 
     /**
@@ -394,12 +520,10 @@ public abstract class Widget {
      * @param color  the color of the circle
      */
     public void drawCircle(int x, int y, float radius, UIColor color) {
-        nvgBeginPath(context.get());
-        nvgCircle(context.get(), x, y, radius);
-        color.tallocNVG(nvgColor -> {
-            nvgFillColor(context.get(), nvgColor);
-        });
-        nvgFill(context.get());
+        nvgBeginPath(getContextID());
+        nvgCircle(getContextID(), x, y, radius);
+        color.tallocNVG(nvgColor -> nvgFillColor(getContextID(), nvgColor));
+        nvgFill(getContextID());
     }
 
     /**
@@ -413,14 +537,12 @@ public abstract class Widget {
      * @param color      the color of the arc segment
      */
     public void drawArc(int x, int y, float radius, float startAngle, float endAngle, UIColor color) {
-        nvgBeginPath(context.get());
-        nvgArc(context.get(), x, y, radius, startAngle, endAngle, NVG_CW);
-        nvgLineTo(context.get(), x, y);
-        nvgClosePath(context.get());
-        color.tallocNVG(nvgColor -> {
-            nvgFillColor(context.get(), nvgColor);
-        });
-        nvgFill(context.get());
+        nvgBeginPath(getContextID());
+        nvgArc(getContextID(), x, y, radius, startAngle, endAngle, NVG_CW);
+        nvgLineTo(getContextID(), x, y);
+        nvgClosePath(getContextID());
+        color.tallocNVG(nvgColor -> nvgFillColor(getContextID(), nvgColor));
+        nvgFill(getContextID());
     }
 
     /**
@@ -435,16 +557,14 @@ public abstract class Widget {
      * @param color      the color of the arc segment
      */
     public void drawArc(int x, int y, float radius, float thickness, float startAngle, float endAngle, UIColor color) {
-        nvgBeginPath(context.get());
-        nvgArc(context.get(), x, y, radius, startAngle, endAngle, NVG_CW);
-        nvgLineTo(context.get(), x + (float) Math.cos(endAngle) * radius, y + (float) Math.sin(endAngle) * radius);
-        nvgArc(context.get(), x, y, radius - thickness, endAngle, startAngle, NVG_CCW);
-        nvgLineTo(context.get(), x + (float) Math.cos(startAngle) * radius, y + (float) Math.sin(startAngle) * radius);
-        nvgClosePath(context.get());
-        color.tallocNVG(nvgColor -> {
-            nvgFillColor(context.get(), nvgColor);
-        });
-        nvgFill(context.get());
+        nvgBeginPath(getContextID());
+        nvgArc(getContextID(), x, y, radius, startAngle, endAngle, NVG_CW);
+        nvgLineTo(getContextID(), x + (float) Math.cos(endAngle) * radius, y + (float) Math.sin(endAngle) * radius);
+        nvgArc(getContextID(), x, y, radius - thickness, endAngle, startAngle, NVG_CCW);
+        nvgLineTo(getContextID(), x + (float) Math.cos(startAngle) * radius, y + (float) Math.sin(startAngle) * radius);
+        nvgClosePath(getContextID());
+        color.tallocNVG(nvgColor -> nvgFillColor(getContextID(), nvgColor));
+        nvgFill(getContextID());
     }
 
     /**
@@ -457,11 +577,11 @@ public abstract class Widget {
      * @param height  the height of the image
      */
     public void drawImage(int imageId, int x, int y, int width, int height) {
-        NVGPaint paint = nvgImagePattern(context.get(), x, y, width, height, 0, imageId, 1, NVGPaint.create());
-        nvgBeginPath(context.get());
-        nvgRect(context.get(), x, y, width, height);
-        nvgFillPaint(context.get(), paint);
-        nvgFill(context.get());
+        NVGPaint paint = nvgImagePattern(getContextID(), x, y, width, height, 0, imageId, 1, NVGPaint.create());
+        nvgBeginPath(getContextID());
+        nvgRect(getContextID(), x, y, width, height);
+        nvgFillPaint(getContextID(), paint);
+        nvgFill(getContextID());
     }
 
     /**
@@ -474,7 +594,7 @@ public abstract class Widget {
      * @param height    the height of the image
      */
     public void drawImage(Path imagePath, int x, int y, int width, int height) {
-        int imageID = nvgCreateImage(context.get(), imagePath.toString(), NVG_IMAGE_NEAREST);
+        int imageID = nvgCreateImage(getContextID(), imagePath.toString(), NVG_IMAGE_NEAREST);
         drawImage(imageID, x, y, width, height);
     }